Adapt helper_cuda.h from cuda samples for error checking

+++ src/matrix.h
@@ -19,6 +19,9 @@
 #ifndef _MATRIX_H_
 #define _MATRIX_H_
 
+#include <cuda.h>
+#include <stdint.h>
+
 #define SQUARE_BLOCK_SIZE 16
 #define SINGLE_BLOCK_SIZE 512
 
@@ -35,7 +38,8 @@
 #define gf uint8_t
 
 // AlignType: the word-alignment length
-typedef unsigned int AlignType;
+// typedef unsigned int AlignType;
+typedef long long AlignType;
 
 // __host__ __device__ uint8_t gf_add(uint8_t a, uint8_t b);
 // __host__ __device__ uint8_t gf_sub(uint8_t a, uint8_t b);
